+++ src/autoresearch/orchestration/orchestrator.py
@@ -24,12 +24,20 @@ The system is designed to be extensible, allowing for custom agents, reasoning m
 and execution strategies.
 """
 
-from typing import List, Dict, Any, Callable, Iterator, TypedDict, cast
+from typing import (
+    List,
+    Dict,
+    Any,
+    Callable,
+    Iterator,
+    TypedDict,
+    cast,
+    ContextManager,
+)
 import os
 import time
 import traceback
 from concurrent.futures import ThreadPoolExecutor
-from contextlib import contextmanager
 import asyncio
 
 from ..agents.registry import AgentFactory, AgentRegistry
@@ -39,6 +47,10 @@ from ..models import QueryResponse
 from ..storage import StorageManager
 from .state import QueryState
 from .metrics import OrchestrationMetrics, record_query
+from .token_utils import (
+    _capture_token_usage as capture_token_usage,
+    _execute_with_adapter as execute_with_adapter,
+)
 from ..logging_utils import get_logger
 from ..tracing import setup_tracing, get_tracer
 from ..errors import OrchestrationError, AgentError, NotFoundError, TimeoutError
@@ -1488,44 +1500,8 @@ class Orchestrator:
     def _execute_with_adapter(
         agent: Any, state: QueryState, config: ConfigModel, adapter: Any
     ) -> Dict[str, Any]:
-        """Execute an agent with a specific adapter.
-
-        This method handles executing an agent with a specific adapter, either by:
-        1. Passing the adapter directly to the execute method if it supports it
-        2. Temporarily setting the adapter in the agent's context
-
-        Args:
-            agent: The agent to execute
-            state: The current query state
-            config: The system configuration
-            adapter: The adapter to use for LLM calls
-
-        Returns:
-            The result of agent execution
-        """
-        # Check if the agent's execute method accepts an adapter parameter
-        import inspect
-
-        sig = inspect.signature(agent.execute)
-
-        if "adapter" in sig.parameters:
-            # Agent supports direct adapter injection
-            return agent.execute(state, config, adapter=adapter)
-        elif hasattr(agent, "set_adapter"):
-            # Agent supports adapter setting via method
-            original_adapter = (
-                agent.get_adapter() if hasattr(agent, "get_adapter") else None
-            )
-            try:
-                agent.set_adapter(adapter)
-                return agent.execute(state, config)
-            finally:
-                if original_adapter is not None:
-                    agent.set_adapter(original_adapter)
-        else:
-            # No adapter injection support, just execute normally
-            # This won't count tokens, but at least it won't break
-            return agent.execute(state, config)
+        """Execute an agent with a specific adapter."""
+        return execute_with_adapter(agent, state, config, adapter)
 
     @staticmethod
     def _rotate_list(items: List[Any], start_idx: int) -> List[Any]:
@@ -1637,57 +1613,11 @@ class Orchestrator:
         return max(0.1, min(1.0, confidence))
 
     @staticmethod
-    @contextmanager
     def _capture_token_usage(
         agent_name: str, metrics: OrchestrationMetrics, config: ConfigModel
-    ) -> Iterator[tuple[dict[str, int], Any]]:
-        """Capture token usage for all LLM calls within the block.
-
-        This method uses the TokenCountingAdapter to count tokens for all LLM calls
-        made within the context manager block. It yields a tuple containing a dictionary
-        with token counts and the wrapped adapter that should be used for LLM calls.
-
-        Args:
-            agent_name: The name of the agent making the LLM calls
-            metrics: The metrics collector to record token usage
-            config: The system configuration containing the LLM backend
-
-        Yields:
-            A tuple containing (token_counter, wrapped_adapter)
-        """
-        from autoresearch.llm.token_counting import count_tokens
-        import autoresearch.llm as llm
-
-        # Get the adapter for the agent using the configured backend
-        backend = config.llm_backend
-        adapter = llm.get_pooled_adapter(backend)
-        token_budget: int | None = getattr(config, "token_budget", None)
-
-        # Use the count_tokens context manager to count tokens
-        # It returns both the token counter and the wrapped adapter
-        with count_tokens(agent_name, adapter, metrics, token_budget) as (
-            token_counter,
-            base_adapter,
-        ):
-            wrapped: Any = base_adapter
-            if token_budget is not None:
-                tb: int = token_budget
-
-                class PromptCompressAdapter:
-                    def __init__(self, inner: Any) -> None:
-                        self.inner = inner
-
-                    def generate(
-                        self, prompt: str, model: str | None = None, **kwargs: Any
-                    ) -> str:
-                        prompt = metrics.compress_prompt_if_needed(prompt, tb)
-                        return self.inner.generate(prompt, model=model, **kwargs)
-
-                    def __getattr__(self, name: str) -> Any:  # pragma: no cover
-                        return getattr(self.inner, name)
-
-                wrapped = PromptCompressAdapter(wrapped)
-            yield token_counter, wrapped
+    ) -> ContextManager[tuple[dict[str, int], Any]]:
+        """Capture token usage for all LLM calls within the block."""
+        return capture_token_usage(agent_name, metrics, config)
 
     # --------------------------------------------------------------
     # Storage helper shortcuts
